fix(render): HiddenMesh tambien viajaba como float
Misma dualidad que se corrigio en BlendMesh: el `a8` de sub_441E00 es un
ENTERO (indice de la malla a ocultar, o -1) y el port lo tipaba `float`, asi
que convivian el bit-pattern (`*(float *)(o + 0x58)`) y el valor numerico
(`-1.0f`, `1.4013e-45f`).

Hoy era inocuo porque las dos codificaciones caen fuera del rango de indices
reales (>= 0) y el test es `i != a8`, o sea las dos dan "ninguna malla
oculta".  Pero era una trampa armada para el proximo call site que pasara un
valor que si discrimine.

El parametro pasa a `int` y se convierten los 44 call sites.  Dos casos no
mecanicos:

  - BMD_SetupRender.cpp: `fVar9` es reuso de registro de Hex-Rays (sirve de
    HiddenMesh en dos caminos y de BlendMeshLight en otro).  Los dos caminos
    que llegan a la llamada del final le asignan `*(float *)(param_2 + 0x58)`,
    asi que se pasa el entero directo y `fVar9` queda como float puro.
  - Render_LegacyBillboards.cpp: `kHiddenMeshAll` era un float construido con
    los bits de -1; ahora es -1 a secas.

+++ src/Render/Entity_AnimationLegacy.cpp
@@ -105,7 +105,7 @@ int __cdecl FUN_004faa70(int param_1, char param_2, int param_3) {
         FUN_00441e00(this_, 0x40,
                      *(float*)(param_1+0x168), *(int *)(param_1+100),
                      *(float*)(param_1+0x68),  *(float*)(param_1+0x6c),
-                     *(float*)(param_1+0x70),  *(float*)(param_1+0x58), 0xffffffff);
+                     *(float*)(param_1+0x70),  *(int *)(param_1+0x58), 0xffffffff);
         FUN_004fa930(param_1, (int)this_);
         _DAT_005597c8 = 1.0f;
     }
